Add files via upload
This Excel file contains the parameters used in my research.
It is important to keep the name of the sheets equal in order to use the Jupiter notebook associated. For those non familiar with Spanish Carga means Load and Precios means prices, so, for better understanding: 
- sheet "CargaTodos" has the different load profiles used 
- sheet "PreciosTodos" has the different costs of energy not supplied for each load 
- sheet "Precios" contains the utility grid prices per month 
- sheet "GridAv" contains the availability profile for each load

There are some others sheets related to PV modules and battery parameters as well as meteorological parameters. and some random hid sheets with resumed information from Puerto Carreño utility grid in 2019 taken from the IPSE web page.