Excel file description before opening it
This Excel file contains the parameters used in my research.
It is important to keep the name of the sheets equal in order to use the Jupiter notebook associated. For those non-familiar with spanish: Carga means Load and Precios means Prices, so, for better understanding: 
- sheet "CargaTodos" has the different load profiles used.
- sheet "PreciosTodos" has the cost of energy not supplied for each load.
- sheet "Precios" contains the utility grid prices per month.
- sheet "GridAv" contains the availability profile for each load.

You can also find here some other sheets related to PV modules and battery parameters as well as meteorological parameters. 
There are hidden sheets with resumed information from Puerto Carreño utility grid in 2019 taken from the IPSE web page that are not really used in the Jupiter Notebook but can be consulted for more details about Puerto Carreño utility grid operation.